+++ src/main/java/com/coastee/server/global/domain/BaseEntityStatus.java
@@ -1,4 +1,4 @@
-package com.coastee.server.global;
+package com.coastee.server.global.domain;
 
 public enum BaseEntityStatus {
     ACTIVE, INACTIVE, DELETED
