+++ src/main/java/com/coastee/server/hashtag/domain/HashTag.java
@@ -1,6 +1,6 @@
 package com.coastee.server.hashtag.domain;
 
-import com.coastee.server.global.BaseEntity;
+import com.coastee.server.global.domain.BaseEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
