+++ src/main/java/com/coastee/server/login/client/OAuthApiClient.java
@@ -5,7 +5,7 @@ import com.coastee.server.login.domain.OAuthUserInfo;
 import com.coastee.server.login.domain.SocialTokens;
 import com.coastee.server.user.domain.SocialType;
 
-import static com.coastee.server.global.Constant.TOKEN_PREFIX;
+import static com.coastee.server.global.domain.Constant.TOKEN_PREFIX;
 
 public interface OAuthApiClient {
     String ACCESS_TOKEN_HEADER = TOKEN_PREFIX;
