+++ src/main/java/com/coastee/server/server/domain/ServerEntry.java
@@ -1,6 +1,6 @@
 package com.coastee.server.server.domain;
 
-import com.coastee.server.global.BaseEntity;
+import com.coastee.server.global.domain.BaseEntity;
 import com.coastee.server.user.domain.User;
 import jakarta.persistence.*;
 import lombok.Builder;
